+++ modules/client/examples/02-addresslistVoting-client/03-create-proposal.ts
@@ -5,7 +5,7 @@ import {
   AddresslistVotingClient,
   Context,
   ContextPlugin,
-  ICreateProposalParams,
+  CreateMajorityVotingProposalParams,
   ProposalCreationSteps,
   ProposalMetadata,
   VoteValues,
@@ -41,7 +41,7 @@ const metadata: ProposalMetadata = {
 
 const ipfsUri = await client.methods.pinMetadata(metadata);
 
-const proposalParams: ICreateProposalParams = {
+const proposalParams: CreateMajorityVotingProposalParams = {
   pluginAddress: "0x1234567890123456789012345678901234567890",
   metadataUri: ipfsUri,
   actions: [],