+++ modules/client/src/client-common/interfaces/plugin.ts
@@ -21,26 +21,27 @@ export enum VoteValues {
 }
 
 // TYPES
-export interface IProposalSettings {
-  /** Float: 0 to 1 */
-  minSupport: number;
-  /** Float: 0 to 1 */
-  minTurnout: number;
-  /** In seconds */
-  duration: number;
-}
 
-export type VotingSettings = {
-  /* default is standard */
-  votingMode?: VotingMode;
+export type MajorityVotingSettingsBase = {
   /** Float between 0 and 1 */
   supportThreshold: number;
   /** Float between 0 and 1 */
   minParticipation: number;
+};
+
+export type MajorityVotingProposalSettings = MajorityVotingSettingsBase & {
+  duration: number;
+};
+export type MajorityVotingSettings = MajorityVotingSettingsBase & {
+  /* default is standard */
+  votingMode?: VotingMode;
+  /* minimum is 3600 */
   minDuration: number;
   /* default is 0 */
   minProposerVotingPower?: bigint;
 };
+
+export type VotingSettings = MajorityVotingSettings
 
 export enum VotingMode {
   STANDARD = "Standard",
@@ -56,19 +57,6 @@ export type ContractVotingSettings = [
   BigNumber, // minProposerVotingPower
 ];
 
-// TODO: turn into a type, and potentially refactor these fields
-export interface ICreateProposalParams {
-  pluginAddress: string;
-  metadataUri: string;
-  actions?: DaoAction[];
-  /** For every action item, denotes whether its execution could fail
-   * without aborting the whole proposal execution */
-  failSafeActions?: Array<boolean>;
-  startDate?: Date;
-  endDate?: Date;
-  executeOnPass?: boolean;
-  creatorVote?: VoteValues;
-}
 
 export type CreateProposalBaseParams = {
   pluginAddress: string;
@@ -78,6 +66,13 @@ export type CreateProposalBaseParams = {
   failSafeActions?: Array<boolean>;
   metadataUri: string;
 };
+
+export type CreateMajorityVotingProposalParams = CreateProposalBaseParams & {
+  startDate?: Date;
+  endDate?: Date;
+  executeOnPass?: boolean;
+  creatorVote?: VoteValues;
+}
 
 export interface IVoteProposalParams {
   vote: VoteValues;