+++ modules/client/src/internal/utils.ts
@@ -1,10 +1,14 @@
 import {
+  AssetBalance,
   ContractPermissionParams,
+  ContractPermissionWithConditionParams,
   DaoDetails,
   DaoListItem,
   DaoMetadata,
   DepositErc20Params,
   DepositEthParams,
+  GrantPermissionWithConditionDecodedParams,
+  GrantPermissionWithConditionParams,
   IGrantPermissionDecodedParams,
   IGrantPermissionParams,
   InstalledPluginListItem,
@@ -18,7 +22,6 @@ import {
   SubgraphPluginTypeMap,
   SubgraphTransferListItem,
   SubgraphTransferType,
-  AssetBalance,
   TokenType,
   Transfer,
   TransferType,
@@ -46,7 +49,7 @@ export function toDaoDetails(
 ): DaoDetails {
   return {
     address: dao.id,
-    ensDomain: dao.subdomain,
+    ensDomain: dao.subdomain + ".dao.eth",
     metadata: {
       name: metadata.name,
       description: metadata.description,
@@ -81,7 +84,7 @@ export function toDaoListItem(
 ): DaoListItem {
   return {
     address: dao.id,
-    ensDomain: dao.subdomain,
+    ensDomain: dao.subdomain + ".dao.eth",
     metadata: {
       name: metadata.name,
       description: metadata.description,
@@ -233,6 +236,18 @@ export function permissionParamsToContract(
 ): ContractPermissionParams {
   return [params.where, params.who, keccak256(toUtf8Bytes(params.permission))];
 }
+export function permissionWithConditionParamsToContract(
+  params: GrantPermissionWithConditionParams,
+): ContractPermissionWithConditionParams {
+  return [
+    ...permissionParamsToContract({
+      who: params.who,
+      where: params.where,
+      permission: params.permission,
+    }),
+    params.condition,
+  ];
+}
 
 export function permissionParamsFromContract(
   result: Result,
@@ -244,6 +259,14 @@ export function permissionParamsFromContract(
     permission: Object.keys(PermissionIds)
       .find((k) => PermissionIds[k] === result[2])
       ?.replace(/_ID$/, "") || "",
+  };
+}
+export function permissionParamsWitConditionFromContract(
+  result: Result,
+): GrantPermissionWithConditionDecodedParams {
+  return {
+    ...permissionParamsFromContract(result),
+    condition: result[3],
   };
 }