+++ modules/client/src/multisig/internal/constants.ts
@@ -4,6 +4,11 @@ import { activeContractsList, Multisig__factory } from "@aragon/core-contracts-e
 
 // TODO: This address needs to be set when the plugin has
 // been published and the ID is known
+// TODO: Find a way around this
+// This shouldnt be hardcoded, but the active contracts need the current network
+// thes means we cannot have it here if we move it to the function where this is used we 
+// still dont have access to the network because is a static function and the context is 
+// not defined. Maybe defining the ens here instead of the address?
 export const MULTISIG_PLUGIN_ID = "0x39e04b9728db34650110f99161AE6d2521D7Cf67";
 
 // TODO update with function names