+++ modules/common/src/errors.ts
@@ -100,7 +100,7 @@ export class InvalidVotingModeError extends Error {
     super("Invalid voting mode");
   }
 }
-export class EnsureAllowanceError extends Error {
+export class UpdateAllowanceError extends Error {
   constructor() {
     super("Could not define a minimum allowance");
   }
@@ -108,5 +108,17 @@ export class EnsureAllowanceError extends Error {
 export class InvalidPrecisionError extends Error {
   constructor() {
     super("Invalid precision, number must be an integer greater than 0");
+  }
+}
+export class FailedDepositError extends Error {
+  constructor() {
+    super("Failed to deposit");
+  }
+}
+export class AmountMismatchError extends Error {
+  constructor(expected: bigint, received: bigint) {
+    super(
+      `Deposited amount mismatch. Expected: ${expected}, received: ${received}`,
+    );
   }
 }